+++ App.js
@@ -8,7 +8,7 @@
 
 import React, { Component } from 'react';
 import Navigator from './app/Navigator';
-import { Platform, StyleSheet, Text, View } from 'react-native';
+// import { Platform, StyleSheet, Text, View } from 'react-native';
 // import Login from './app/Screens/Login';
 // import StockList from './app/Screens/StockList';
 // import Register from './app/Screens/Register';