+++ app/Components/TickerCard.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import {
-  Text, View, TouchableOpacity, Alert, StyleSheet, Dimensions, 
+  Text, View, TouchableOpacity, Alert, StyleSheet, Dimensions,
 } from 'react-native';
 import Ionicon from 'react-native-vector-icons/Ionicons';
 
@@ -11,7 +11,7 @@ export default class TickerCard extends Component {
   }
 
   render() {
-    let { id, name, price } = this.props;
+    const { id, name, price } = this.props;
 
     // No price means it's just the list for Add Tickers
     if (price === 'ADD') {
@@ -30,11 +30,11 @@ export default class TickerCard extends Component {
     else {
       return (
         <View style={styles.card}>
-          <View style={styles.left}>
+          <View>
             <Text style={styles.idText}>{id}</Text>
             <Text style={styles.nameText}>{name}</Text>
           </View>
-          <View style={styles.right}>
+          <View>
             <Text style={styles.priceText}>${price}</Text>
           </View>
         </View>
@@ -75,5 +75,6 @@ const styles = StyleSheet.create({
   },
   priceText: {
     fontSize: 20,
+    color: 'black',
   },
 });