+++ NEWS.md
@@ -1,6 +1,7 @@
 # NEWS
 
 ## May 24, 2019
+- Added ability to view detailed ticker information displaying current price, support and resistance levels, and strengths.
 - Fixed bug where user's already added stocks were being displayed in Add Symbol page. 
 
 ## May 16, 2019